+++ generators/app/templates/src/StarterKit/DataAccess/Context/EntityContext.mongo.cs
@@ -1,7 +1,6 @@
-using StarterKit.DataAccess.Context;
 using StarterKit.DataAccess.Options;
 
-namespace StarterKit.DataAccess
+namespace StarterKit.DataAccess.Context
 {
   public class EntityContextMongo : ContextBase
   {
